requirements.txt
pytesseract
langchain_core
langchain_groq
pillow
pymupdf
numpy

